requirements.txt
streamlit
pdfplumber
python-docx
nltk
scikit-learn
numpy
